Completely offloaded rendering and window handling to Lua.
The question is how much we offload to Lua. On estimates, Lua can be up to 100x as slow as C. Maybe all rendering, complicated math, physics, and neural networks can be written in C. What then does that leave for Lua? Where do we draw the line between C and Lua?